Allow Dash to be opened from the "Quick Help" popup without replacing the popup entirely. (refs #14)
The "Enable Dash Quick Help" menu item has been replaced by a "Style" menu
which gives three options: disable Dash altogether, replace the Quick Help popup,
or replace the documentation viewer. The default option is, as before, to replace
the Quick Help popup.

When the third, "Replace Reference" option is selected, the plugin allows
the Quick Help popup to be shown, but then intercepts taps on links in that popup
and shows Dash instead of Xcode's documentation viewer, using
the last path component of the link URL as the string for which to search.
This works both for "reference" links (at the bottom of the quick help popup)
and other links.